fix: 식단 이미지 삭제 시, 이미지 값 유무 체크 로직 추가

+++ src/main/java/BE_Elixir/Elixir/domain/dietLog/service/DietLogService.java
@@ -99,8 +99,9 @@ public class DietLogService {
         }
 
         // S3 버킷에서 이미지 삭제
-        s3Service.deleteS3(dietLog.getImageUrl(), "diet_log");
-
+        if (dietLog.getImageUrl() != null) {
+            s3Service.deleteS3(dietLog.getImageUrl(), "diet_log");
+        }
         // 식단 삭제
         dietLogRepository.delete(dietLog);
     }
